Support configuring Gitaly's transaction in specs
Gitaly has a configuration flag that enables support
for transactions with ACID properties. With the flag
enabled, Gitaly runs each RPC in a transaction leading
to improved reliability.

As a preliminary step towards rolling out transactions,
this commit adds support for configuring transactions in
the test setup by setting `GITALY_TRANSACTIONS_ENABLED=true`
environment variable.

In the CI, Gitaly's configuration is generated ahead of the
running the rspec jobs that conditionally run with
transactions enabled. If we're running in CI, transactions
are by default disabled. We generate an additional configuration
file with transactions enabled and run Gitaly with it if
transactions are enabled. This is similar to the setup we
currently have with Praefect as it generates a config with
and without database being enabled in the setup-test-env job.