Release raising permissions of group APIs to owner for package settings
We identified a discrepancy between the UI & the API Group Owners / administrators permission for package and registries settings. We addressed this discrepancy and raised the permissions of the group-level APIs to only the owner, but the changes were behind a feature flag. In this MR, we cleanup the feature flag and release the fix.

Changelog: fixed