Fix a dependency proxy upload premature interruption in workhorse
Clients could close the connection before the dependency
proxy logic had time to finish the upload of the upstream
resource. This would cancel the upload.

The upload request will now not re-use the context of the
client request as-is to avoid this situation. Instead, it will
use a new context that will follow the first context and provide a
grace period to let the upload request complete.

Changelog: fixed