Use more meaningful audience in ID token examples
The audience of the JWT should be explicit—matching the party it was intended to be given to / used by. 
In these examples, the token is created to attest identity to the Vault server, so the audience is the Vault server. 
Using the opposite party (gitlab instance) as the audience would prevent Vault from validating
if a provided JWT was issued for usage in Vault server or if it was issued for a different service and 
is being co-opted to authenticate with Vault.