Added initial state request on merge request dashboard
This creates the component and logic to send the GraphQL requests on the merge
request dashboard.
With this there is an array passed from the backend that contains
what states to display, what queries to send and what variables to send with them.
For now this is to keep the configuration easy to update and change, but it
opens us up to user based customisation in the future if we want to go that route.