Lift trigger_request association up to Ci::Processable
Both `Ci::Build` and `Ci::Bridge` belong to `trigger_request` and share
variable builder methods that depend on it. We recently added
`CI_TRIGGER_SHORT_TOKEN`, but the necessary method delegation was only
added on `Ci::Build`, which caused a problem when when a `Ci::Bridge`
was triggered.

The immediate fix was to add the missing delegation, but in a follow-up
discussion we decided that the association and similar delegations
should probably be on `Ci::Processable` instead.

See https://gitlab.com/gitlab-org/gitlab/-/issues/457571