Skip requested changes check when setting MWCP
Before this change, when MR is blocked because reviewer requested
changes, the MR can't be set to MWCP because the requested changes
check isn't skipped in that case. This results to `merge` action
returning a `status: "failed"` response.

To fix this behavior, when MR is set to MWCP and
`additional_merge_when_checks_ready` FF is enabled, we skip the
requested changes check. This allows the MR to be set to MWCP even
if the check isn't successful yet.

No changelog since this is dependent on 3 feature flags:
- `merge_when_checks_pass`
- `additional_merge_when_checks_ready`
- `mr_reviewer_requests_changes`