Use MIN_SCHEMA_GITLAB_VERSION in keep for finalizing migrations
The OverdueFinalizeBackgroundMigration  keep is a script that is run in
CI for generating merge requests to
[finalize batched background migrations](https://docs.gitlab.com/ee/development/database/batched_background_migrations.html#finalize-a-batched-background-migration).

Per our documentation we can only finalize a batched background
migration added before the last "required stop" which is the minimum
versin someone must upgrade to before they can upgrade to this version
of the code.

We were hardcoding that version in the keep before but now we have
MIN_SCHEMA_GITLAB_VERSION which is exactly for this purpose we don't
need to hardcode it anymore. We also changed from `<` to `<=` now that
this version refers to the version of the last required stop rather than
the first version after it.