Improve CI rules:exists pattern matching performance with cache
In rules:exists, if a file path is a pattern (including `*`), the path
pattern is iterated through every single file in the repository via
`File.fnmatch?`. This is an expensive operation so it is limited to
`MAX_PATTERN_COMPARISONS` (10k for now) per pattern.

We want to increase this limit
(https://gitlab.com/gitlab-org/gitlab/-/issues/227632) but we need to
improve (https://gitlab.com/gitlab-org/gitlab/-/issues/351593)
the performance problem first.

Here, we are adding a request-store caching per glob. With this, a glob
is iterated only once. We can accept the fact that files do not change
during a request or a worker processing.

This is behind a feature flag `ci_rules_exists_pattern_matches_cache`.