Implement basic version of automated code review
This introduces a `ai_review` quick action which will execute a
LLM completion called `ReviewMergeRequest`.

This will make a request to AI for each diff hunk and create a note
for each diff hunk reviewed.

Diff files that are not diffable and deleted will be excluded from
the review.

This is behind the `ai_review_merge_request` feature flag so no
changelog.