Apply rate limit to project hook test API endpoint
Testing a webhook performs an HTTP request to the webhook receiver
and waits for the response. With our authenticated API rate limit being
2,000/m this makes us vulnerable to potentially tie up API node
resources. Especially when the external service is slow to respond.

This change applies a rate limit to the endpoint of 3 requests per
minute, per project webhook.

It also adjusts the urgency to `:low`, to adjust the target duration to
`5s`.

https://gitlab.com/gitlab-org/gitlab/-/issues/25329#note_1868556805

Changelog: changed