Remove confidential attribute from Note
After the announcement in 15.5 that we rename `confidential` to
`internal` for notes, we now remove the `confidential`
attribute on the GraphQL API for `Notes`.

It's no longer possible to create notes using `confidential` and it's
also no longer possible to query `confidential` on notes.
As a replacement, use the `internal` field.

Changelog: removed