Add CTA to purchase more seats when importing project members fails
When block seat overages is enabled, group owners cannot add more
members to a group than the number of seats in the subscription.
When attempting to import project members from one project to another,
the import will fail with an error message if there are not enough
seats left in the subscription.
Add a CTA to purchase more seats to the error message.
The CTA appears only for the group owner.
Block seat overages is behind a feature flag.