Rescue from validation errors when uploading terraform modules
If a terraform module uploaded with invalid name, a validation error will be raised. However, the error will not be gracefully rescued and returned to the user. It will cause a 500 internal server error. In this MR, we rescue from such errors and return them to the user with the appropriate status code and error message.

Changelog: fixed