Updated count on merge request badge in navigation
This updates the count for the merge request button in the navigation bar when
the `merge_request_dashboard` feature flag is enabled.
When the `merge_request_dashboard` feature flag is enabled the count includes
merge requests that are waiting for my review, and merge requests I am an assignee
of that a reviewer has requested changes.