Fix conflict between canceling methods
Active Record enum adds methods like `canceling!` for each
status on a Stage. We also defined `canceling!` as a state machine
event which conflicts with the enum method. Here we rename
the event machine method so we can take advantage of the
state machine event controls and still transtion to `canceling`.
Most of are current event names are not the exact status name.