Swap the running order of subheadings
Move the newer "auto-merge" subheading above "merge when checks pass"
so we can start building in place in future MRs.

Expand the description of the checks

Creates a paraphrased (this is important) list of checks that are
handled by this feature. I think it important we keep something in
both the user and API docs, but that means they shouldn't be 1:1
cognates of each other.

Swap the running order of subheadings

Move the newer "auto-merge" subheading above "merge when checks pass"
so we can start building in place in future MRs.

Expand the description of the checks

Creates a paraphrased (this is important) list of checks that are
handled by this feature. I think it important we keep something in
both the user and API docs, but that means they shouldn't be 1:1
cognates of each other.

rewrite page for future auto-merge state

rename page to auto_merge

add redirect page

add redirect