Support rules:exists subkeys `paths`, `project`, and `ref`
Adds the ability to specify the context in which `rules:exists` searches
for the given file paths. The subkeys `paths`, `project`, and `ref` can
be specified.  If `project` is specified but not `ref`, the default
`ref` is the project's HEAD. If `project` is not specified, the defaults
are as follows:

For `job:rules:exists`, the default context is the project/ref in
which the pipeline is running.

For `include:rules:exists`, the default context is the project/ref
of the file in which the `include` is declared.

Feature flag `ci_support_rules_exists_paths_and_project` removed.

Changelog: added