Cells ADR 005: Cells 1.0 internal customers
What
---
Create an ADR for Cells 1.0 being internal customers only

Why
---
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150629 we
updated the existing documentation specifying that Cells 1.0 is for
internal customers only, but didn't track this as a decision.

Since this is a big change we want to capture this as a decision.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>