Add decision about initial cell sizes
What
---
Create a new ADR documenting the discussion around cell sizing happening
in https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/2838#accepted-proposal

Why
---
Right now it's an issue, and it's hard to discover.
Putting it in the blueprint makes it easier to link too,
and will generate a record if we want to update it as well.

Signed-off-by: Steve Xuereb <sxuereb@gitlab.com>