Add batched migration to remove namespace from OS components
When we first started ingesting OS components from Trivy SBOMs, we did
not remove the namepaces from the component names. This meant that we
did not match advisories because a component name in the advisory data
was listed without the namespace prefix.

For example, we'd get the cURL component for an Alpine Linux image
as `pkg:apk/alpine/curl@<version>`, which we'd ultimately parse as
`alpine/curl`. In this example, the advisories for cURL would have a
name like `curl` instead of `alpine/curl`. Thus, we'd never get a match
on the advisories and components when performing advisory scans.

The PURL specification dictates that `alpine/curl` is the correct name,
but in our case we choose to deviate from this because store the
operating system metadata as an SBOM property instead. All this
considered, the most pragmatic solution here, and thus the one taken, is
to migrate all of our components to drop the namespace/os prefixes
instead of re-exporting all of our advisories from the GitLab Package
Metadata Database (a much larger change).

Changelog: changed