Add gql option to ciCatalogResources to sort by usage
Adds a GraphQl option to the 'ciCatalogResources'
endpoint to sort by last_30_day_usage_count.
Usage is defined as the number of unique projects that
have used a component from the catalog resource in the
last 30 days, using 'include:component'.  For now the
values are all 0 but they will be populated in
https://gitlab.com/gitlab-org/gitlab/-/issues/452545.

Also added an index on the usage column to improve
query performance when sorting by last_30_day_usage_count.

Changelog: added