Fix 500 errors when viewing empty CI artifacts image
When generating a ZIP metadata artifact, Workhorse will omit the
`size` field if the file is empty due to the way `omitempty` works in
Go. Attempting to view the file in the artifacts browser will result
in a 500 error as a result.

Fix this by defaulting to a size of 0 if `size` doesn't exist. Update
the test artifact ZIP and metadata to test this issue.

When spec/fixtures/ci_build_artifacts.zip was updated a few years ago,
the corresponding metadata was not updated. This commit fixes the
non-UTF-8 directory case and regenerates the metadata with the latest
`gitlab-zip-metadata` version.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/460378

Changelog: fixed