gitlab-zip-metadata: Ensure 0-byte files have a size field
As mentioned in https://pkg.go.dev/encoding/json#Marshal:

The "omitempty" option specifies that the field should be omitted from
the encoding if the field has an empty value, defined as false, 0, a
nil pointer, a nil interface value, and any empty array, slice, map,
or string.

As a result, for 0-byte files, the JSON output of the ZIP metadata
omits the `crc` and `size` fields. Since we always expect this to be
generated, drop the `omitempty` entries from `crc`, `size`, and
`zipped` fields.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/460378

Changelog: fixed