Release avoiding copy operation in terraform modules upload
When uploading a terraform module, we put the file in a temp location on object store, then copy the file from the temp location to a permanent one. In this MR, we avoid this copy operation and upload the file to its final location from the beginning. This helps save time & cost in packages upload.

Changelog: performance