Allow auto merge on created and blocked pipelines
Enables the flag `auto_merge_when_incomplete_pipeline_succeeds` by
default, resulting in the following changes:

Firstly, it fixes a bug where if merge trains were enabled, AND
"Pipelines must succeed" was disabled, AND the head pipeline was
blocked, then the default merge action was an immediate merge as opposed
to "Add to merge train".

Secondly, when "Pipelines must succeed" is enabled, it makes it so that
merge requests with a blocked head pipeline (i.e. with status `manual`
and `scheduled`) can be "Set to auto-merge" when the pipeline succeeds.
This is only made available when "Pipelines must succeed" is enabled,
because otherwise the pipeline does not block the merge, and immediate
merges are well established there as the default merge action.

- Feature issue: https://gitlab.com/gitlab-org/gitlab/-/issues/334281
- Rollout issue: https://gitlab.com/gitlab-org/gitlab/-/issues/454162

Changelog: fixed