Adds specs for gitlab backup Restore subcommand
This adds basic specs for the new classes added in gitlab-backup-cli.
These changes do not go out of the way to stub and verify the
actual code that does the restoration actions, as they are
basic wrappers around classes defined in ::Backup (lib/backup).
Adding those specs is left to a later issue in the project.