Allow creating top-level groups in an organization using `POST /groups`
Currently, to create top-level groups in an organization we use the
`POST /api/v4/groups` API. But this API doesn't work for top-level
groups. Therefore, we're adding the JSON response type to the controller
so that we can make an async request to the controller to create a
top-level group.