Prevent reporting mismatch for deleted records
It could happen that while we compare an epic and the work item record
to validate that they are in sync, that these records get deleted in the
database. This could lead to false warnings for the mismatched
attributes. Especially for attributes that are part of the epic record
but are stored in a separate table on the work item.

For example:
- We load the epic and the work item record
- We start comparing attributes
- The epic and work item record gets deleted on the database
- We compare epic.start_date (which is still in memory) and query
  work_item_dates_sources which returns nothing.
- We report a mismatch, although the data no longer exists.

As a first step of preventing these false reports, we now only log
mismatches when the record still exists in the database at the time of
logging.