Replace custom inject function with common injectVueAppBreadcrumbs
The removed `injectVueRouterIntoBreadcrumbs` was implemented before
the reusable `injectVueAppBreadcrumbs` became available.

Judging by how similar the code was, probably the former was even the
inspiration for the later. But it never got migrated to it. Maybe
because it was not fully compatible, as the former required an inject,
which the later doesn't allow to provide.

I now updated the only usage of it, to not provide/inject the
`groupPath`, but read it from the DOM where we always "provide" it.