Introduce new helpers for style tags and paths
We want to leverage vite directly for providing style assets during
development. In order to have more fine-grained control over what is
imported, we introduce `universal_stylesheet_link_tag` and
`universal_path_to_stylesheet` which are wrappers around
`stylesheet_link_tag` and `stylesheet_path` respectively.

In a follow-up MR we will extend the newly introduced wrappers to use
vite if vite is enabled:
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152044

Note that there are two direct usages of `stylesheet_path` with
`lazy_bundles/cropper.css` but given that it is a CSS file and not SCSS,
we don't support it yet.