Remove duplicated index from abuse_reports
## What does this MR do and why?

Remove duplicated index from `abuse_reports` table.

### It removes:

Drop `index_abuse_reports_on_user_id` as it's already covered by:
| Index | Columns |
| ----- | ------ |
| `idx_abuse_reports_user_id_status_and_category` | `user_id asc`, `status asc`, `category asc` |

It is possible that this MR will still need some changes to drop the index from the database.
Currently, the `gitlab-housekeeper` is not always capable of removing all references, so you must check the diff and pipeline failures to confirm if there are any issues.
Ensure that the index exists in the production database by checking Joe Bot trough https://console.postgres.ai/gitlab.
If the index was already removed or if the index it's being removed in another merge request, consider closing this merge request.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
using the Keeps::RemoveDuplicatedIndexes keep.

To provide feedback on your experience with `gitlab-housekeeper` please comment in
<https://gitlab.com/gitlab-org/gitlab/-/issues/442003>.

Changelog: other