Prevent WebIDE from crashing on message
The full pre-receive secret detection message was crashing in the
WebIDE. This is an attempt to get it to crash less. It will still crash
when there are a lot of secrets detected. I got it to crash at 8
secrets. We will be implementing a more permanent solution with
https://gitlab.com/gitlab-org/gitlab/-/issues/456548

Changelog: fixed