Release changing npm package upload from sync to async
Due to the upload of npm packages happening synchronously we face some limitations and unable to parse the package tarball. That's why we merged a change to make the upload happen async behind a feature flag. In this MR, we cleanup the feature flag and release the change.

Changelog: changed