Make Sidekiq Web UI tests less flaky
One of the Sidekiq Web UI test attempts to test that the latency shown
in the queues page is 0.0. However, if there is some slowness on the
runner the latency could tick up to a second, causing the test to
fail.

Avoid this flakiness by stubbing `Time.now` so that we guarantee that
the latency will always remain at 0.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/460842