Add mutation for clearing project runner cache
Currently clearing a project's runner cache is only possible via the UI.
Users have requested that it be available via an API as well. This
commit adds a new RunnerCacheClear GraphQL mutation that can be used to
clear a project's runner cache. It requires that the current user have
`admin_pipeline` permissions on the project, which is the same
permission used in the UI.

Changelog: added
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/390684