Enable FF_NETWORK_PER_BUILD=true for all vendored gems
In https://gitlab.com/gitlab-org/gitlab/-/issues/460842, we discovered
that the legacy mechanism for linking Docker containers via the
`--link` argument assigns both IPv4 and IPv6 addresses. Since the
experimental IPv6 option is not enabled in Docker, talking to
network services incurs about a 1-second delay as the network
falls back to IPv4.

To avoid this, we should switch on `FF_NETWORK_PER_BUILD=true`,
which is the default setting in our private runners. This
ensures only IPv4 addresses are used.