Add reassigned_by_user to import_source_users
This MR adds a field to import_source_users to track the user that
initiated the user mapping assignment. We probably can't assume that it
will always be the owner of the destination group, so I think it makes
sense to track it separately.

This field is not currently used, but will be when the logic for
assigning destination users is added.

Changelog: added