Updates link on security developer template
The `security/developer.md` file on release/docs was renamed to
`security/engineer.md` as part of efforts to adjust the release
documentation to use the new patch release terminology. This commit
adjusts the issue template to use the new link.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/20184