Add controllers for user contribution mapping
This MR introduces the basic controller functionality to support the
assignee-facing side of https://gitlab.com/gitlab-org/gitlab/-/issues/443555.

There is more work to be done on the interface, but this MR enables the
basic flow of accepting/rejecting assignment, and introduces the flag to
be used across the feature.