Explicitly set Omnibus and CNG Ruby version in CI
This commit adds `OMNIBUS_RUBY_VERSION` and `CNG_RUBY_VERSION` to
define which versions are used in the Omnibus and CNG projects.

Previously the Ruby versions used in the child pipelines for building
Omnibus and CNG was determined based on the image used in the
respective CI jobs. When we bumped the CI image to 3.1.5 in
https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/803,
this quietly added a requirement to Omnibus that Ruby 3.1.5 had to be
supported. At the same time, the `build-cng` image used Ruby 3.1.4, so
https://gitlab.com/gitlab-org/build/CNG-mirror/-/jobs/6821715942
started to fail once CNG stopped building Ruby 3.1.4.

While the latter case needed
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/1828, we can
at least avoid the first problem by explicitly setting the version
needed. In an ideal world, these variables would be determined
automatically from the upstream projects. However, setting these
variables avoids surprise failures and opens the possibility up for
automation to update these values.