Make default admin email not a known value
Currently, when GitLab is first installed and set up regardless of
method, the default admin user `root` likely has the known email address
`admin@example.com` . This makes it easier to exploit if there is an
email-based vulnerability.

This MR adds some slight randomization to the initial admin email
address, so it is hard to guess.

Changelog: changed