Change Jest sharding strategy in CI
This changes our Jest sharding strategy from round-robin to path sorted.

With the original round-robin, 1/12th of all frontend specs across the
codebase were run in a given shard. That meant lots of unrelated specs
were running in the same shard, and related specs were spread across
shards.

Our [theory] was that this increased resource usage in jobs, both in
terms of memory (disparate subtrees of the codebase being imported,
which added up to roughly most of the codebase) and compute (transpiling
common source files).

So, we looked at changing our sharding and sorting strategies, to put
related specs in the same job.

In [practice], it turned out not to make much of a difference! Still,
there is a benefit to the change.

Now, there's more consistency/reliability about which spec runs on which
shard. Before, the addition of a new spec would move, on average, half
of all specs to a difference shard, compare to now, fewer than 12 specs
would move.

[theory]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150196#note_1880916097
[practice]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152536

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/461073.