Remove saml_reload feature flag
`saml_reload` has been enable globably in GitLab.com. We deemed the
feature stable and remove the feature flag that toggle it.

Users will see a modal when their SAML session has expired. This prevent
users from losing comments or other action that may not work as expected
because of the expiration of the session.

Changelog: changed