Warm up Jest cache before running tests in CI
This is an attempt to reduce the frequency with which Jest jobs fail due
to timeouts in CI.

Experiments conducted in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150196 show that
Jest spec run time in CI is highly variable and inconsistent. This can
also be seen in the latest Jest speed report:
https://gitlab-org.gitlab.io/frontend/playground/jest-speed-reporter/.

One way to mitigate this is to reduce the amount of work Jest needs to
do in each job by preparing an import cache for Jest and providing it to
each job. This cache should prevent *some* unnecessary IO/compute on
these jobs, mostly around transpilation of files, done by at least
`vue2-jest`, `babel`, and our GraphQL transformer.

This seems to reduce the worst cases of slow specs, which should help
reduce the frequency of timeouts. There doesn't appear to be much of
a benefit for the majority/long tail of tests, however.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/457179.

Co-Authored-By: Lukas Eipert <leipert@gitlab.com>
Co-Authored-By: Paul Slaughter <pslaughter@gitlab.com>