Add CI jobs with Gitaly transactions
Add rspec jobs that run with Gitaly's transactions enabled. We
add new jobs as we're not yet running transcations in production
and thus want to ensure we test the production code paths as
usual. The new jobs ensure we don't regress anymore while we are
preparing transactions for a rollout.