Document Gitaly storage breaking change
With 17.0 Gitaly will no longer allow multiple storages to
use the same path. As we previously recommended tthat admins do this in the docs,
we expect this to have some impacts.

Add this to the breaking change document and list a workaround
needed if any projects were in a duplicate storage.