Fix error when calling GQL ciConfig endpoint with include:component
Currently, when requesting the includes->extra field from the
`ciConfig` GraphQl endpoint, it returns an internal server response
when the config content contains `include:component`. This happens
because the component file class returns a Project object in the
`extra` field. And serializing an object is forbidden.

This MR fixes this problem by moving the Project object and
related attributes out of the `extra` field and into a new metadata
field named `component`. This new field is not exposed by the API
and there does not raise the serialization error.

Changelog: fixed