Introduce component usage aggregator and Redis cursor
Introduces new component usage aggregator and cursor classes
to be used later in a service scheduled by cron job. The objective is
to run a regular worker that aggregates the CI component usage data
in the background, avoiding long running queries.

The Usages::Aggregator class iterates through the target scope in
batches. For each target ID, it aggregates the component usage
count in batches for the given window. It maintains a Redis cursor
so that the counter can resume from it left off on each run and
avoid unnecessary reprocessing. We collect the count in Rails
because the SQL query `COUNT(DISTINCT(*))` is not performant when
the dataset is large. This class should be run with an exclusive
lease for each unique cursor Redis key.