Dont download whole GCP objects in clean up task
For GCP object storage provider, we had to fetch the object to be able
to track the last generation attribute value before we delete it. This
is needed by the rollback rake task if ever.

We encountered memory usage issues when executing this because the call
to `artifacts_directory.files.get` would download the whole object
including the content which ended up problematic for large artifact
objects.

The fix here is to just use `artifacts_directory.files.metadata`, which
is only available for `fog-google`, so that it will only download the
object metadata.