Adds mlModelDelete mutation
To better align with our usage of GraphQL, mlModelDestroy needs to be
changed to mlModelDelete and the return type needs to be the model. The
first step is to add new mutation mlModelDelete so that we can then
change the frontend, otherwise the frontend can be deployed beforehand
and will try to access an mlModelDelete that doesn't exist yes.

Deleting mlModelDestroy and changing the frontend to call mlModelDelete
are followups.