Revisit the Summary section of CI Time Decay doc
With this change we make the Summary self sufficient as
an executive summary. Previously it was giving a lot of 
historical context without describing what we are doing.

The Summary should stay releveant with the approaches
we decide to take with the architectural design doc.

It also descibes the 4 main strategies we intend to apply.